% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtubeAnalytics_objects.R
\name{BatchReport}
\alias{BatchReport}
\title{YouTube Analytics API Objects 
Retrieves your YouTube Analytics data.}
\usage{
BatchReport(BatchReport.outputs = NULL, BatchReport.timeSpan = NULL,
  id = NULL, outputs = NULL, reportId = NULL, timeSpan = NULL,
  timeUpdated = NULL)
}
\arguments{
\item{BatchReport.outputs}{The \link{BatchReport.outputs} object or list of objects}

\item{BatchReport.timeSpan}{The \link{BatchReport.timeSpan} object or list of objects}

\item{id}{The ID that YouTube assigns and uses to uniquely identify the report}

\item{outputs}{Report outputs}

\item{reportId}{The ID of the the report definition}

\item{timeSpan}{Period included in the report}

\item{timeUpdated}{The time when the report was updated}
}
\value{
BatchReport object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2016-09-03 23:51:26
filename: /Users/mark/dev/R/autoGoogleAPI/googleyoutubeAnalyticsv1.auto/R/youtubeAnalytics_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
BatchReport Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Contains single batchReport resource.
}
\seealso{
Other BatchReport functions: \code{\link{BatchReport.outputs}},
  \code{\link{BatchReport.timeSpan}}
}

