% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtubeAnalytics_objects.R
\name{BatchReportDefinition}
\alias{BatchReportDefinition}
\title{BatchReportDefinition Object}
\usage{
BatchReportDefinition(id = NULL, name = NULL, status = NULL,
  type = NULL)
}
\arguments{
\item{id}{The ID that YouTube assigns and uses to uniquely identify the report definition}

\item{name}{Name of the report definition}

\item{status}{Status of the report definition}

\item{type}{Type of the report definition}
}
\value{
BatchReportDefinition object
}
\description{
BatchReportDefinition Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Contains single batchReportDefinition resource.
}

