% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtubeAnalytics_objects.R
\name{BatchReportDefinitionList}
\alias{BatchReportDefinitionList}
\title{BatchReportDefinitionList Object}
\usage{
BatchReportDefinitionList(items = NULL)
}
\arguments{
\item{items}{A list of batchReportDefinition resources that match the request criteria}
}
\value{
BatchReportDefinitionList object
}
\description{
BatchReportDefinitionList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A paginated list of batchReportDefinition resources returned in response to a youtubeAnalytics.batchReportDefinitions.list request.
}

