% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtubeAnalytics_functions.R
\name{batchReports.list}
\alias{batchReports.list}
\title{Retrieves a list of processed batch reports.}
\usage{
batchReports.list(batchReportDefinitionId, onBehalfOfContentOwner)
}
\arguments{
\item{batchReportDefinitionId}{The batchReportDefinitionId parameter specifies the ID of the batch reportort definition for which you are retrieving reports}

\item{onBehalfOfContentOwner}{The onBehalfOfContentOwner parameter identifies the content owner that the user is acting on behalf of}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/yt-analytics-monetary.readonly
\item https://www.googleapis.com/auth/yt-analytics.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/yt-analytics-monetary.readonly, https://www.googleapis.com/auth/yt-analytics.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{http://developers.google.com/youtube/analytics/}{Google Documentation}
}

