% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtubeAnalytics_objects.R
\name{Group.contentDetails}
\alias{Group.contentDetails}
\title{Group.contentDetails Object}
\usage{
Group.contentDetails(itemCount = NULL, itemType = NULL)
}
\arguments{
\item{itemCount}{No description}

\item{itemType}{No description}
}
\value{
Group.contentDetails object
}
\description{
Group.contentDetails Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Group functions: \code{\link{Group.snippet}},
  \code{\link{Group}}, \code{\link{groups.insert}},
  \code{\link{groups.update}}
}

