% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtubeAnalytics_objects.R
\name{Group}
\alias{Group}
\title{Group Object}
\usage{
Group(Group.contentDetails = NULL, Group.snippet = NULL,
  contentDetails = NULL, etag = NULL, id = NULL, snippet = NULL)
}
\arguments{
\item{Group.contentDetails}{The \link{Group.contentDetails} object or list of objects}

\item{Group.snippet}{The \link{Group.snippet} object or list of objects}

\item{contentDetails}{No description}

\item{etag}{No description}

\item{id}{No description}

\item{snippet}{No description}
}
\value{
Group object
}
\description{
Group Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Group functions: \code{\link{Group.contentDetails}},
  \code{\link{Group.snippet}}, \code{\link{groups.insert}},
  \code{\link{groups.update}}
}

