% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtubeAnalytics_objects.R
\name{Group.snippet}
\alias{Group.snippet}
\title{Group.snippet Object}
\usage{
Group.snippet(publishedAt = NULL, title = NULL)
}
\arguments{
\item{publishedAt}{No description}

\item{title}{No description}
}
\value{
Group.snippet object
}
\description{
Group.snippet Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Group functions: \code{\link{Group.contentDetails}},
  \code{\link{Group}}, \code{\link{groups.insert}},
  \code{\link{groups.update}}
}

