% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtubeAnalytics_objects.R
\name{GroupItem}
\alias{GroupItem}
\title{GroupItem Object}
\usage{
GroupItem(GroupItem.resource = NULL, etag = NULL, groupId = NULL,
  id = NULL, resource = NULL)
}
\arguments{
\item{GroupItem.resource}{The \link{GroupItem.resource} object or list of objects}

\item{etag}{No description}

\item{groupId}{No description}

\item{id}{No description}

\item{resource}{No description}
}
\value{
GroupItem object
}
\description{
GroupItem Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other GroupItem functions: \code{\link{GroupItem.resource}},
  \code{\link{groupItems.insert}}
}

