% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtubeAnalytics_objects.R
\name{GroupItem.resource}
\alias{GroupItem.resource}
\title{GroupItem.resource Object}
\usage{
GroupItem.resource(id = NULL)
}
\arguments{
\item{id}{No description}
}
\value{
GroupItem.resource object
}
\description{
GroupItem.resource Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other GroupItem functions: \code{\link{GroupItem}},
  \code{\link{groupItems.insert}}
}

