% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtubeAnalytics_objects.R
\name{ResultTable}
\alias{ResultTable}
\title{ResultTable Object}
\usage{
ResultTable(ResultTable.columnHeaders = NULL, columnHeaders = NULL,
  rows = NULL)
}
\arguments{
\item{ResultTable.columnHeaders}{The \link{ResultTable.columnHeaders} object or list of objects}

\item{columnHeaders}{This value specifies information about the data returned in the rows fields}

\item{rows}{The list contains all rows of the result table}
}
\value{
ResultTable object
}
\description{
ResultTable Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Contains a single result table. The table is returned as an array of rows that contain the values for the cells of the table. Depending on the metric or dimension, the cell can contain a string (video ID, country code) or a number (number of views or number of likes).
}
\seealso{
Other ResultTable functions: \code{\link{ResultTable.columnHeaders}}
}

