% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtubeAnalytics_functions.R
\docType{package}
\name{youtubeAnalytics_googleAuthR}
\alias{youtubeAnalytics_googleAuthR}
\alias{youtubeAnalytics_googleAuthR-package}
\title{YouTube Analytics API
Retrieves your YouTube Analytics data.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2016-09-03 23:51:25
filename: /Users/mark/dev/R/autoGoogleAPI/googleyoutubeAnalyticsv1.auto/R/youtubeAnalytics_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item https://www.googleapis.com/auth/youtube
\item https://www.googleapis.com/auth/youtube.readonly
\item https://www.googleapis.com/auth/youtubepartner
\item https://www.googleapis.com/auth/yt-analytics-monetary.readonly
\item https://www.googleapis.com/auth/yt-analytics.readonly
}
}

